chore(stores): move all types to types.ts
